Added support for searching strangers by language and informing being mobile.
The Omegle class has now two new parameters, language and mobile.
By default, the value of language is set to "en" (English) and
mobile false. These can be modified the same way as userAgent and
host as explained in the "Further Asides" section of readme.md.

I also cleaned up the constructor a bit.